float implementati

+++ meter/meter.h
@@ -10,7 +10,7 @@
 #define ARRAY_SIZE 1000
 
 typedef struct __attribute__((packed)) amp_value {
-  uint16_t current; 
+  float current; 
   uint16_t timestamp;
 } amp_value;
